Move LocalizationTranslationFilter
Move LocalizationTranslationFilter from CensusDataFilter to specific
filter.
